Split nodes into nodes and utils for easier management
